[FEAT] fixes and test improvement

+++ Tryitter/Controllers/postController.cs
@@ -73,8 +73,15 @@ public class PostController : ControllerBase
   }
 
   [HttpDelete("{id}")]
-  public IActionResult Delete(int id)
+  public async Task<IActionResult> DeleteAsync(int id)
   {
+    var post = await _repository.GetById(id);
+
+    if (post == null)
+    {
+      return BadRequest("Algo deu errado!");
+    }
+
     _repository.Delete(id);
 
     return NoContent();
